requirements.txt
torch
Flask
opencv-python
Pillow
flask_socketio
ultralytics
